common/python/requirements.txt
azure-storage-queue
azure-storage-blob